Running better!
I updated how the simulation is started so that the first agents are less
uniform. I also added randomness to how many agents can exist at any one
time. This had improved the realism of the simulation. Next is to start
experimentation which I plan to do using an automated system.