Ran the first experimental test.
The simulation is running great and as expected. However I do need to
change when the json files are loaded which is what I think is slowing
down the simulation. The test run covering all cases at 10 per took
about an hour. The resulting files were too large to upload to github so
they are being saved on a separate medium.