Changed when agents are reassigned when org or com is priority.
I tried to compensate for the impact of motivation on output by allowing
organizations to change their members' jobs if a resource falls below a
threshold. The threshold was determined by running trials that tracked
the balance of resources. These runs were also used to modulate resource
availability so that output better matches needs.